requirements.txt
flask
optimum-intel[openvino]>=1.12.1
transformers>=4.35.0
torch>=2.0.0
pillow
scipy
sounddevice
psutil
soundfile
librosa>=0.10.0
ffmpeg-python
audioread
